test: add ImageIO backend
The current ImageMagick version for Julia >= 1.3 is a bit broken for PNG files
because the built ImageMagick_jll is broken.
https://github.com/JuliaIO/ImageMagick.jl/issues/206

This commit work arounds this bug by adding ImageIO dependency.
Because ImageIO has higher priority in the FileIO registry, it will
use the PNGFiles to load the data instead of using the broken
ImageMagick.